+++ build.sh
@@ -1,11 +1,14 @@
 # Create the dist directory if needed
 if [[ ! -d dist ]]
-	then mkdir dist
+	then mkdir dist dist/xslt dist/config
 fi
 # Likewise, create the package dir
 if [[ ! -d package ]]
 	then mkdir package
 fi
+
+# Get the current version
+VERSION=`grep -o ' packageVersion \"\(.*\)\"' version.ent | awk '{print $2}' | sed 's/"//g'`
 
 # Make sure to use the PRODUCTION entities
 UMBOFF="umbraco \"IGNORE\""
@@ -30,8 +33,8 @@ sed -i "" "s/\&amp;\(.*\);/\&\1;/" package/_CalendarHelper.xslt
 sed -i "" "s/\&amp;\(.*\);/\&\1;/" package/_MediaHelper.xslt
 
 # Copy configs
-cp mediahelpers/cropping-config.xml package/cropping-config.xml
-cp calendarhelper/calendar-config.xml package/calendar-config.xml
+cp mediahelpers/CroppingSettings.config package/CroppingSettings.config
+cp calendarhelper/CalendarSettings.config package/CalendarSettings.config
 
 # Copy default templates
 cp templates/Use*.xslt package/
@@ -40,16 +43,16 @@ cp templates/Use*.xslt package/
 xsltproc --novalid --xinclude --output package/package.xml lib/freezeEntities.xslt package.xml
 
 # Build the ZIP file 
-zip -j dist/XSLTHelpers package/* -x \*.DS_Store
+zip -j "dist/XSLTHelpers-$VERSION.zip" package/* -x \*.DS_Store
 
 # Copy the release XSLT into the dist dir for upgraders
-cp package/_PaginationHelper.xslt dist/_PaginationHelper.xslt
-cp package/_NavigationHelper.xslt dist/_NavigationHelper.xslt
-cp package/_GroupingHelper.xslt dist/_GroupingHelper.xslt
-cp package/_CalendarHelper.xslt dist/_CalendarHelper.xslt
-cp package/_MediaHelper.xslt dist/_MediaHelper.xslt
-cp package/cropping-config.xml dist/cropping-config.xml
-cp package/calendar-config.xml dist/calendar-config.xml
+cp package/_PaginationHelper.xslt dist/xslt/_PaginationHelper.xslt
+cp package/_NavigationHelper.xslt dist/xslt/_NavigationHelper.xslt
+cp package/_GroupingHelper.xslt dist/xslt/_GroupingHelper.xslt
+cp package/_CalendarHelper.xslt dist/xslt/_CalendarHelper.xslt
+cp package/_MediaHelper.xslt dist/xslt/_MediaHelper.xslt
+cp package/CroppingSettings.config dist/config/CroppingSettings.config
+cp package/CalendarSettings.config dist/config/CalendarSettings.config
 
 # Go back to DEVELOPMENT versions again
 sed -i "" "s/$UMBON/$UMBOFF/" */entities.ent