+++ Release Notes.md
@@ -1,15 +1,19 @@
 # Release Notes
 
+* v0.8.3: Calendar Helper updated to 1.1, Media Helpers updated to 1.3
 * v0.8.2: Media Helpers updated to 1.2
 * v0.8.1: Pagination Helper updated to 1.1
 * v0.8: Media Helpers updated to 1.1
 
 ## Calendar Helper
 
+* v1.1 Fix for issue #12, - implement issue #11
+
 ## Grouping Helper
 
 ## Media Helpers
 
+* v1.3 Move (and rename) cropping config file to config folder - issue #11
 * v1.2 Adding `retinafy` option
 
 ## Navigation Helper